Generate the normal distribution of a radiocarbon age

Generate the normal distribution of a radiocarbon age

Usage

c14_age_norm(age, error, resolution = 1, sigma = 5)

Arguments

age

Radiocarbon age(s) in years.

error

Error(s) associated with the radiocarbon age in years.

resolution

Desired resolution in years.

sigma

Desired sigma range.

Value

A vector of probability density with years as names. If age or error has length >= 1, a list of these vectors.

Details

age and error are recycled to a common length.

Examples

c14_age_norm(10000, 30)
#>         9850         9851         9852         9853         9854         9855 
#> 4.955732e-08 5.851254e-08 6.900928e-08 8.129869e-08 9.567028e-08 1.124574e-07 
#>         9856         9857         9858         9859         9860         9861 
#> 1.320433e-07 1.548682e-07 1.814369e-07 2.123275e-07 2.482015e-07 2.898145e-07 
#>         9862         9863         9864         9865         9866         9867 
#> 3.380284e-07 3.938254e-07 4.583231e-07 5.327914e-07 6.186715e-07 7.175967e-07 
#>         9868         9869         9870         9871         9872         9873 
#> 8.314157e-07 9.622180e-07 1.112362e-06 1.284507e-06 1.481644e-06 1.707139e-06 
#>         9874         9875         9876         9877         9878         9879 
#> 1.964769e-06 2.258767e-06 2.593874e-06 2.975389e-06 3.409228e-06 3.901986e-06 
#>         9880         9881         9882         9883         9884         9885 
#> 4.461008e-06 5.094454e-06 5.811386e-06 6.621849e-06 7.536961e-06 8.569012e-06 
#>         9886         9887         9888         9889         9890         9891 
#> 9.731564e-06 1.103957e-05 1.250947e-05 1.415934e-05 1.600902e-05 1.808023e-05 
#>         9892         9893         9894         9895         9896         9897 
#> 2.039673e-05 2.298448e-05 2.587177e-05 2.908942e-05 3.267093e-05 3.665265e-05 
#>         9898         9899         9900         9901         9902         9903 
#> 4.107397e-05 4.597751e-05 5.140930e-05 5.741896e-05 6.405993e-05 7.138962e-05 
#>         9904         9905         9906         9907         9908         9909 
#> 7.946961e-05 8.836587e-05 9.814890e-05 1.088940e-04 1.206812e-04 1.335958e-04 
#>         9910         9911         9912         9913         9914         9915 
#> 1.477283e-04 1.631743e-04 1.800352e-04 1.984177e-04 2.184344e-04 2.402033e-04 
#>         9916         9917         9918         9919         9920         9921 
#> 2.638484e-04 2.894992e-04 3.172909e-04 3.473645e-04 3.798662e-04 4.149477e-04 
#>         9922         9923         9924         9925         9926         9927 
#> 4.527656e-04 4.934817e-04 5.372619e-04 5.842767e-04 6.347000e-04 6.887092e-04 
#>         9928         9929         9930         9931         9932         9933 
#> 7.464843e-04 8.082077e-04 8.740630e-04 9.442346e-04 1.018907e-03 1.098264e-03 
#>         9934         9935         9936         9937         9938         9939 
#> 1.182486e-03 1.271754e-03 1.366242e-03 1.466120e-03 1.571552e-03 1.682696e-03 
#>         9940         9941         9942         9943         9944         9945 
#> 1.799699e-03 1.922700e-03 2.051827e-03 2.187194e-03 2.328903e-03 2.477039e-03 
#>         9946         9947         9948         9949         9950         9951 
#> 2.631672e-03 2.792853e-03 2.960615e-03 3.134969e-03 3.315905e-03 3.503388e-03 
#>         9952         9953         9954         9955         9956         9957 
#> 3.697361e-03 3.897741e-03 4.104417e-03 4.317253e-03 4.536083e-03 4.760712e-03 
#>         9958         9959         9960         9961         9962         9963 
#> 4.990916e-03 5.226441e-03 5.467002e-03 5.712286e-03 5.961947e-03 6.215610e-03 
#>         9964         9965         9966         9967         9968         9969 
#> 6.472868e-03 6.733290e-03 6.996410e-03 7.261739e-03 7.528761e-03 7.796933e-03 
#>         9970         9971         9972         9973         9974         9975 
#> 8.065691e-03 8.334447e-03 8.602594e-03 8.869508e-03 9.134549e-03 9.397063e-03 
#>         9976         9977         9978         9979         9980         9981 
#> 9.656385e-03 9.911845e-03 1.016276e-02 1.040846e-02 1.064827e-02 1.088150e-02 
#>         9982         9983         9984         9985         9986         9987 
#> 1.110749e-02 1.132558e-02 1.153513e-02 1.173551e-02 1.192611e-02 1.210635e-02 
#>         9988         9989         9990         9991         9992         9993 
#> 1.227567e-02 1.243353e-02 1.257944e-02 1.271293e-02 1.283356e-02 1.294096e-02 
#>         9994         9995         9996         9997         9998         9999 
#> 1.303476e-02 1.311466e-02 1.318039e-02 1.323175e-02 1.326856e-02 1.329069e-02 
#>        10000        10001        10002        10003        10004        10005 
#> 1.329808e-02 1.329069e-02 1.326856e-02 1.323175e-02 1.318039e-02 1.311466e-02 
#>        10006        10007        10008        10009        10010        10011 
#> 1.303476e-02 1.294096e-02 1.283356e-02 1.271293e-02 1.257944e-02 1.243353e-02 
#>        10012        10013        10014        10015        10016        10017 
#> 1.227567e-02 1.210635e-02 1.192611e-02 1.173551e-02 1.153513e-02 1.132558e-02 
#>        10018        10019        10020        10021        10022        10023 
#> 1.110749e-02 1.088150e-02 1.064827e-02 1.040846e-02 1.016276e-02 9.911845e-03 
#>        10024        10025        10026        10027        10028        10029 
#> 9.656385e-03 9.397063e-03 9.134549e-03 8.869508e-03 8.602594e-03 8.334447e-03 
#>        10030        10031        10032        10033        10034        10035 
#> 8.065691e-03 7.796933e-03 7.528761e-03 7.261739e-03 6.996410e-03 6.733290e-03 
#>        10036        10037        10038        10039        10040        10041 
#> 6.472868e-03 6.215610e-03 5.961947e-03 5.712286e-03 5.467002e-03 5.226441e-03 
#>        10042        10043        10044        10045        10046        10047 
#> 4.990916e-03 4.760712e-03 4.536083e-03 4.317253e-03 4.104417e-03 3.897741e-03 
#>        10048        10049        10050        10051        10052        10053 
#> 3.697361e-03 3.503388e-03 3.315905e-03 3.134969e-03 2.960615e-03 2.792853e-03 
#>        10054        10055        10056        10057        10058        10059 
#> 2.631672e-03 2.477039e-03 2.328903e-03 2.187194e-03 2.051827e-03 1.922700e-03 
#>        10060        10061        10062        10063        10064        10065 
#> 1.799699e-03 1.682696e-03 1.571552e-03 1.466120e-03 1.366242e-03 1.271754e-03 
#>        10066        10067        10068        10069        10070        10071 
#> 1.182486e-03 1.098264e-03 1.018907e-03 9.442346e-04 8.740630e-04 8.082077e-04 
#>        10072        10073        10074        10075        10076        10077 
#> 7.464843e-04 6.887092e-04 6.347000e-04 5.842767e-04 5.372619e-04 4.934817e-04 
#>        10078        10079        10080        10081        10082        10083 
#> 4.527656e-04 4.149477e-04 3.798662e-04 3.473645e-04 3.172909e-04 2.894992e-04 
#>        10084        10085        10086        10087        10088        10089 
#> 2.638484e-04 2.402033e-04 2.184344e-04 1.984177e-04 1.800352e-04 1.631743e-04 
#>        10090        10091        10092        10093        10094        10095 
#> 1.477283e-04 1.335958e-04 1.206812e-04 1.088940e-04 9.814890e-05 8.836587e-05 
#>        10096        10097        10098        10099        10100        10101 
#> 7.946961e-05 7.138962e-05 6.405993e-05 5.741896e-05 5.140930e-05 4.597751e-05 
#>        10102        10103        10104        10105        10106        10107 
#> 4.107397e-05 3.665265e-05 3.267093e-05 2.908942e-05 2.587177e-05 2.298448e-05 
#>        10108        10109        10110        10111        10112        10113 
#> 2.039673e-05 1.808023e-05 1.600902e-05 1.415934e-05 1.250947e-05 1.103957e-05 
#>        10114        10115        10116        10117        10118        10119 
#> 9.731564e-06 8.569012e-06 7.536961e-06 6.621849e-06 5.811386e-06 5.094454e-06 
#>        10120        10121        10122        10123        10124        10125 
#> 4.461008e-06 3.901986e-06 3.409228e-06 2.975389e-06 2.593874e-06 2.258767e-06 
#>        10126        10127        10128        10129        10130        10131 
#> 1.964769e-06 1.707139e-06 1.481644e-06 1.284507e-06 1.112362e-06 9.622180e-07 
#>        10132        10133        10134        10135        10136        10137 
#> 8.314157e-07 7.175967e-07 6.186715e-07 5.327914e-07 4.583231e-07 3.938254e-07 
#>        10138        10139        10140        10141        10142        10143 
#> 3.380284e-07 2.898145e-07 2.482015e-07 2.123275e-07 1.814369e-07 1.548682e-07 
#>        10144        10145        10146        10147        10148        10149 
#> 1.320433e-07 1.124574e-07 9.567028e-08 8.129869e-08 6.900928e-08 5.851254e-08 
#>        10150 
#> 4.955732e-08