Radiocarbon calibration curve class

The c14_curve class represents a radiocarbon calibration curve, which estimates the calendar ages corresponding to a set of radiocarbon measurements. Usually calibration is performed with either one of the standard curves included with the package (see c14_curves) or read in from a file with read_14c().

c14_curve() constructs a a new calibration curve using values in 14C years, optionally with their associated d14C values.

c14_fcurve() constructs a curve using fraction modern values (F14C or pMC), typically used for "post-bomb" calibration.

c14_curve(cal_age, c14_age, c14_error, d14c = NA, d14c_error = NA)

c14_fcurve(cal_age, f14c, f14c_error)

Arguments

cal_age
Vector of calendar ages. If cal_age is an object of class era::yr, its epoch will be respected, otherwise it is assumed to be "cal BP".

c14_age
Vector of uncalibrated radiocarbon ages.

c14_error
Vector of errors (sigma values) associated with c14_age.

d14c
Vector of Delta-14C values (optional).

d14c_error
Vector of errors (sigma values) associated with d14c (optional).

f14c
Vector of fraction modern (F14C or pMC) values.

f14c_error
Vector of errors (sigma values) associated with f14c.

Value

Object of class c14_curve.

References

Stenström KE, Skog G, Georgiadou E, Genberg J, Johansson A (2011). “A guide to radiocarbon units and calculations.” Technical Report LUNFD6(NFFR-3111)/1-17/(2011), Lund University, Nuclear Physics.

See also