Radiocarbon calibration curves

Radiocarbon calibration requires a calibration curve: an estimate of the true calendar age associated with a radiocarbon measurement. The following curves are included with the c14 package:

IntCal20

IntCal Northern Hemisphere curve, 0–55,000 cal BP Reimer2020c14.

SHCal20

IntCal Southern Hemisphere curve, 0–55,000 cal BP Hogg2020c14.

Marine20

IntCal marine curve, 0–55,000 cal BP Heaton2020c14.

IntCal13, IntCal09, IntCal04, IntCal98

Previous IntCal Northern Hemisphere curves Reimer2013,Reimer2009,Reimer2004,Stuiver1998c14. Superseded by IntCal20.

SHCal13, SHCal04

Previous IntCal Southern Hemisphere curves Hogg2013,McCormac2004c14. Superseded by SHCal20.

Marine13, Marine09, Marine04, Marine98

Previous IntCal marine curves Reimer2013,Reimer2009,Hughen2004,Stuiver1998c14. Superseded by Marine20.

Generally, it is recommended to use the latest curve from the IntCal series: IntCal20 for samples from the northern hemisphere; SHCal20 for samples from the southern hemisphere; or Marine20 for samples that obtained their 14C in a marine environment.

Usage

IntCal20

SHCal20

Marine20

IntCal13

SHCal13

Marine13

IntCal09

Marine09

IntCal04

SHCal04

Marine04

IntCal98

Marine98

Format

A c14_curve object with the following fields.

For standard curves:

cal_age

Calendar ages with associated era (era::yr class).

c14_age

Uncalibrated radiocarbon ages.

c14_error

Errors (sigma values) associated with c14_age.

d14c

Delta-14C values.

d14c_error

Errors (sigma values) associated with d14c.

For post-bomb curves:

cal_age

Calendar ages with their associated era (era::yr class).

f14c

Fraction modern (F14C or pMC) values.

f14c_error

Errors (sigma values) associated with f14c.

An object of class c14_curve_14c (inherits from c14_curve, vctrs_rcrd, vctrs_vctr) of length 9501.

An object of class c14_curve_14c (inherits from c14_curve, vctrs_rcrd, vctrs_vctr) of length 9501.

An object of class c14_curve_14c (inherits from c14_curve, vctrs_rcrd, vctrs_vctr) of length 5501.

An object of class c14_curve_14c (inherits from c14_curve, vctrs_rcrd, vctrs_vctr) of length 5141.

An object of class c14_curve_14c (inherits from c14_curve, vctrs_rcrd, vctrs_vctr) of length 5141.

An object of class c14_curve_14c (inherits from c14_curve, vctrs_rcrd, vctrs_vctr) of length 4801.

An object of class c14_curve_14c (inherits from c14_curve, vctrs_rcrd, vctrs_vctr) of length 3522.

An object of class c14_curve_14c (inherits from c14_curve, vctrs_rcrd, vctrs_vctr) of length 3651.

An object of class c14_curve_14c (inherits from c14_curve, vctrs_rcrd, vctrs_vctr) of length 3302.

An object of class c14_curve_14c (inherits from c14_curve, vctrs_rcrd, vctrs_vctr) of length 2482.

An object of class c14_curve_14c (inherits from c14_curve, vctrs_rcrd, vctrs_vctr) of length 3301.

An object of class c14_curve_14c (inherits from c14_curve, vctrs_rcrd, vctrs_vctr) of length 1538.

An object of class c14_curve_14c (inherits from c14_curve, vctrs_rcrd, vctrs_vctr) of length 1569.

Source

http://intcal.org/curves/ and http://calib.org/CALIBomb/

References