Radiocarbon unit calculations

Functions for calculating basic units used in radiocarbon measurements.

c14_age() calculates the conventional radiocarbon age (CRA) from a fraction modern measurement.

c14_f14c() reverse-calculates the fractionation-corrected fraction modern value (F^14CF14C or pMpM) of a radiocarbon age.

Usage

c14_age(x, decay = c14::c14_decay_libby)

c14_f14c(x, decay = c14::c14_decay_libby)

Arguments

x

For c14_age(), a vector of fraction modern (F^14CF14C) measurements. For c14_f14c(), a vector of conventional radiocarbon ages.

decay

Decay constant. The default is the Libby constant (c14_decay_libby), which is the standard for calculating conventional radiocarbon ages. Use c14_decay_cambridge for the Cambridge constant, or a single numeric for other values.

Value

Vector the same length as x.

Details

c14_age() calculates the conventional radiocarbon age, tt, as defined by Stuiver1977;textualc14:

t = -1F^14Ct = -1/l * ln(F^14C)

c14_f14c() implements the inverse of this function:

F^14C = e^- tF14C = e^(-lt)

The decay constant conventionally used for calculating radiocarbon ages is the Libby decay constant, _L=8033^-1lL = 1/8033. An alternative is the Cambridge decay constant, _C=8267^-1lC = 1/8267 Stenstrom2011c14.

Reported radiocarbon ages are usually rounded based on the magnitude of the error Stuiver1977c14. For this reason, reverse-calculating fraction modern from a radiocarbon age is unlikely to return the precise original measurement of the sample.

Where available, fraction modern is the recommended measurement for calibration Bronk_Ramsey2008c14.

References

Examples

c14_age(0.9239)
#> [1] 0.999885
c14_f14c(636)
#> [1] -51854.61